Recent Posts with Category Exclusion Widget
In this post, I'm going to share an extremely useful widget for your
Wordpress projects. I'm sure that you've heard about the Wordpress
recent posts widget. It creates links to your recent posts in your
sidebar so your visitors can reach the other blog posts you've written
easily. It's a great tool to create traffic within your blog.

But the widget we have in the Wordpress core can be too simplistic for
most of the well maintained blogs and it needs modifications. For
instance, you may want to show only the recent posts belonging to a
specific category, or you may want to exclude some categories. Yes,
there are many great plugins to achieve this functionality but a plugin
for a simple task like this seems like an overkill. This code will help
you to achieve those goals without installing any plugins. Just copy and
paste it to your functions.php file and you'll get this when you go to
the recent posts widget